requirements.txt
pandas>=2.0.0
prophet>=1.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

